V0.6 Style Settings Start
Rewrote the color scheme to work with Style Settings. For now, there are some theme options (it's pretty bare bones, I know) to change between the default-like styling, and some experimental color schemes. The new schemes should be 90% done, the last changes will be color tweaks to determine the right contrast. Fonts have also been slightly changed. The new interface and editor font is Rubik, a rounded font that shouldn't be too light or heavy; the new title font is Zoho Puvi, an awesome (and super satisfying) font with a bunch of cool quirks (did you know the angle of the letters is based off of the earths axial tilt?).